+++ es/core/connect.js
@@ -1,12 +1,17 @@
 import register from './register';
 import { MODULE_DEFAULT } from '../support/constant';
-export default function (ccClassKey, stateToPropMapping, module) {
+export default function (ccClassKey, stateToPropMapping, isPropStateModuleMode, module) {
+  if (isPropStateModuleMode === void 0) {
+    isPropStateModuleMode = false;
+  }
+
   if (module === void 0) {
     module = MODULE_DEFAULT;
   }
 
   return register(ccClassKey, {
     module: module,
-    stateToPropMapping: stateToPropMapping
+    stateToPropMapping: stateToPropMapping,
+    isPropStateModuleMode: isPropStateModuleMode
   });
 }