+++ es/core/startup.js
@@ -236,6 +236,7 @@ function bindComputedToCcContext(computed, isModuleMode) {
 
 function executeInitializer(isModuleMode, store, init) {
   var stateHandler = helper.getStateHandlerForInit;
+  if (init === undefined) return;
 
   if (!isModuleMode) {
     if (isPlainJsonObject(init)) {
@@ -246,20 +247,24 @@ function executeInitializer(isModuleMode, store, init) {
         if (includeDefaultModule) {
           var defaultInit = init[MODULE_DEFAULT];
 
-          if (typeof defaultInit !== 'function') {
-            throw new Error('init.$$default value must be a function when cc startup in nonModuleMode!');
-          } else {
-            defaultInit(stateHandler(MODULE_DEFAULT));
+          if (defaultInit) {
+            if (typeof defaultInit !== 'function') {
+              throw new Error('init.$$default value must be a function when cc startup in nonModuleMode!');
+            } else {
+              defaultInit(stateHandler(MODULE_DEFAULT));
+            }
           }
         }
 
         if (includeGlobalModule) {
           var globalInit = init[MODULE_GLOBAL];
 
-          if (typeof globalInit !== 'function') {
-            throw new Error('init.$$global value must be a function when cc startup in nonModuleMode!');
-          } else {
-            globalInit(stateHandler(MODULE_GLOBAL));
+          if (globalInit) {
+            if (typeof globalInit !== 'function') {
+              throw new Error('init.$$global value must be a function when cc startup in nonModuleMode!');
+            } else {
+              globalInit(stateHandler(MODULE_GLOBAL));
+            }
           }
         }
       } else {
@@ -284,7 +289,10 @@ function executeInitializer(isModuleMode, store, init) {
       }
 
       var initFn = init[moduleName];
-      initFn(stateHandler(moduleName));
+
+      if (initFn) {
+        initFn(stateHandler(moduleName));
+      }
     });
   }
 }