+++ src/core/connect.js
@@ -1,6 +1,6 @@
 import register from './register';
 import { MODULE_DEFAULT } from '../support/constant';
 
-export default function (ccClassKey, stateToPropMapping, module = MODULE_DEFAULT) {
-  return register(ccClassKey, { module, stateToPropMapping });
+export default function (ccClassKey, stateToPropMapping, isPropStateModuleMode = false, module = MODULE_DEFAULT) {
+  return register(ccClassKey, { module, stateToPropMapping, isPropStateModuleMode });
 }